Refactor progress utilities and types
