+++ components/progress/Progress.types.ts
@@ -1,3 +1,5 @@
+import type { ReactNode } from "react";
+
 export type TrendPoint = {
   x: Date; // Local date
   y: number; // value
@@ -80,4 +82,76 @@ export type CardioProgressSnapshot = {
   series: Record<CardioFocus, CardioSeriesResponse>;
   kpis: CardioKpi[];
   workouts: Record<string, CardioWorkoutSummary[]>;
+};
+
+export type TrendSummary = {
+  icon: string;
+  color: string;
+  colorActive: string;
+  text: string;
+  delta: number;
+};
+
+export type DateInput = string | number | Date | undefined | null;
+
+export type BucketTotals = {
+  minutes: number;
+  distanceKm: number;
+  calories: number;
+  steps: number;
+  heartRateSum: number;
+  heartRateCount: number;
+};
+
+export type Bucket = {
+  start: Date;
+  end: Date;
+  totals: BucketTotals;
+  workouts: CardioWorkoutSummary[];
+};
+
+export type AggregatedData = {
+  range: TimeRange;
+  current: Bucket[];
+  previous: Bucket[];
+};
+
+export type AggregatedTotals = {
+  calories?: number;
+  distance?: number;
+  steps?: number;
+  time?: number;
+};
+
+export type CardioWeekHistoryWorkout = {
+  id: number | string;
+  name: string;
+  source?: string;
+  duration?: ReactNode;
+  time?: string;
+  calories?: ReactNode;
+  distance?: ReactNode;
+  steps?: ReactNode;
+  exercises?: number;
+  sets?: number;
+  rounds?: number;
+  volume?: ReactNode;
+  personalRecords?: number;
+  type?: string;
+  start?: Date;
+  end?: Date;
+};
+
+export type CardioWeekHistoryDay = {
+  key: string;
+  label?: string;
+  weekIndex: number;
+  dateLabel: string;
+  dailyTotals: {
+    calories?: ReactNode;
+    time?: ReactNode;
+    distance?: ReactNode;
+    steps?: ReactNode;
+  };
+  workouts: CardioWeekHistoryWorkout[];
 };
